Add blog post case study for TrustMesh and HCS standards (#170)
* Add case study for TrustMesh and HCS standards

Added a detailed case study on TrustMesh, a Hedera Africa Hackathon prototype submission. The doc is outlining its origins, core concepts, system architecture, and future proposals.

* Revise TrustMesh case study with new content and structure

Updated the case study for TrustMesh to reflect new title, authors, and content structure. Added details on HCS-21 and its implications for decentralized trust signaling.

* Rename trustmesh to 2025-10-18-hcs-21-trustmesh-scend.md

* Update TrustMesh logo with a link

* Modify slug and enhance TrustMesh case study content

Updated slug for the case study and added a new section about a day in the life of TrustMesh on campus.

* Update and rename 2025-10-18-hcs-21-trustmesh-scend.md to 2025-11-4-hcs-21-trustmesh-scend.md

* Update 2025-11-4-hcs-21-trustmesh-scend.md

* Refactor TrustMesh case study for HCS-X proposal

Updated the slug, title, and tags for the TrustMesh case study. Revised sections to reflect the proposed HCS standard and clarified the purpose and next steps.

---------

Co-authored-by: Michael Kantor <6068672+kantorcodes@users.noreply.github.com>